Autogenerating types for my local queries like it's 2020 :robot:
